+++ CHANGELOG.md
@@ -5,7 +5,7 @@ BUG FIXES:
  * consul/connect: Fixed a bug where gateway proxy connection default timeout not set [[GH-9851](https://github.com/hashicorp/nomad/pull/9851)]
  * scheduler: Fixed a bug where shared ports were not persisted during inplace updates for service jobs. [[GH-9830](https://github.com/hashicorp/nomad/issues/9830)]
 
-## 1.0.2 (January 14, 2020)
+## 1.0.2 (January 14, 2021)
 
 IMPROVEMENTS:
  * artifact: Added support for virtual host style AWS S3 paths. [[GH-9050](https://github.com/hashicorp/nomad/issues/9050)]